+++ tools/HeaderForge/include/Parser.hpp
@@ -7,21 +7,29 @@
 #ifndef PARSER_HPP
 #define PARSER_HPP
 
+#include <set>
 #include <string>
 #include <vector>
 
 namespace Engine::Serialization {
   /// Represents a member variable with its fully-qualified type name and identifier.
   struct Variable {
-    std::string type; ///< Fully qualified type name (including namespace, templates, pointers).
-    std::string name; ///< Variable name as declared in the class.
+    /// Fully qualified type name (including namespace, templates, pointers).
+    std::string type;
+    /// Variable name as declared in the class.
+    std::string name;
   };
 
   /// Represents a record (class/struct) with its name, location, and serializable fields.
   struct Record {
-    std::string name; ///< Fully qualified name of the class/struct.
-    std::string filePath; ///< File path where this class is defined.
-    std::vector<Variable> fields; ///< List of serializable member variables.
+    /// Fully qualified name of the class/struct.
+    std::string name;
+    /// File path where this class is defined.
+    std::string filePath;
+    /// List of serializable member variables.
+    std::vector<Variable> fields;
+    /// List of all template parameters of this record
+    std::vector<std::string> templateParameters;
   };
 
   /**
@@ -34,9 +42,13 @@ namespace Engine::Serialization {
        * Parse the given header file and return a list of Records representing
        * serializable classes and their fields.
        * @param headerPath Path to the header file to parse.
+       * @param records All the extracted records
+       * @param usages All the extracted record template usages
        * @return Vector of Record structures with class and field info.
        */
-      static std::vector<Record> ParseHeader(const std::string &headerPath);
+      static void ParseHeader(
+        const std::string &headerPath, std::vector<Record> &records, std::set<std::string> &usages
+      );
 
       /**
        * Utility function to print the collected records and their fields to standard output.
@@ -44,7 +56,9 @@ namespace Engine::Serialization {
        */
       static void PrintRecords(const std::vector<Record> &records);
     private:
+      /// @returns The path to the clang resource directory
       static std::string getClangResourceDir();
+      /// @returns The path to the macOS SDK
       static std::string getMacOSSDKPath();
   };
 } // namespace Engine::Serialization